test(kernels): draw read-raf fixture indices from each table's valid domain
The instruction read-raf parity fixtures fed uniform random lookup indices
to present tables. Tables with constrained operand domains (shift/rotate
bitmask operands, lane-mask operands) only match materialize_entry on the
valid domain, so the LaneExtractS rows (sampled as the last enum variant)
broke the reference kernel's round checks.

Adopt the in-domain convention: random_lookup_index becomes an instance
method exposed through LookupTableKind behind a test-utils feature, the
bitmask generator moves to tables::index_gen, and fixture_rows draws each
table row's index from its table's domain (edge-index rows carry no table).

Also regenerate the inline-expansion golden fixture (BLAKE3 713 -> 701
rows rides the LW/SW lane recipes) and the main expansion parity fixture
(102 memory-family hashes; word-arithmetic hashes stay at main's values).
