+++ crates/jolt-lookup-tables/Cargo.toml
@@ -13,6 +13,10 @@ workspace = true
 
 [features]
 prover-abi-tests = []
+# Exposes per-table in-domain random lookup index generation to downstream
+# test fixtures (tables with constrained operand domains override the
+# uniform default).
+test-utils = ["dep:rand"]
 field-inline = [
   "jolt-prover-legacy/field-inline",
   "jolt-riscv/field-inline",
@@ -22,6 +26,7 @@ field-inline = [
 [dependencies]
 jolt-field = { workspace = true }
 jolt-riscv = { workspace = true, features = ["serialization"] }
+rand = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 strum = { workspace = true, features = ["derive"] }
