+++ jolt-inlines/fixtures/fixtures/registered_inline_expand_parity_hashes.jsonl
@@ -14,10 +14,10 @@
 {"name":"BLAKE2_INLINE::compressed","inline_name":"BLAKE2_INLINE","opcode":11,"funct3":0,"funct7":2,"extension":"Blake2","row_count":1072,"output_sha256":"820c30f8e1ea03705e922c61493c5618e55756f1400c63254ff0f4c651be7998"}
 {"name":"BLAKE2_INLINE::rd_zero","inline_name":"BLAKE2_INLINE","opcode":11,"funct3":0,"funct7":2,"extension":"Blake2","row_count":1072,"output_sha256":"44e02c0f1b34e23c3a3c6d7b35cadd49b94cbc5b0fcf17b8d0498f4f928c7071"}
 {"name":"BLAKE2_INLINE::aliased_operands","inline_name":"BLAKE2_INLINE","opcode":11,"funct3":0,"funct7":2,"extension":"Blake2","row_count":1072,"output_sha256":"458f48b5b2b178869609dee609034dd67fcdf33edab2720b333a45972f62af5e"}
-{"name":"BLAKE3_INLINE::normal","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":713,"output_sha256":"087877e498522d6952d2e5cb9a6eb625fa0ae6e044e9482d298c00f5b2da3c55"}
-{"name":"BLAKE3_INLINE::compressed","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":713,"output_sha256":"afc65a727faa3639ae9f4c6c4efe1de6f8c73867df85c352406c398800e32f9d"}
-{"name":"BLAKE3_INLINE::rd_zero","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":713,"output_sha256":"7f95b5bd4a470962d401699f0998e49c686e5cf26336be88afb9a187e0fe3d89"}
-{"name":"BLAKE3_INLINE::aliased_operands","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":713,"output_sha256":"dd016d78424161d43748c7fe75e90bb67e67513120ac6fe1449af26d212f4ceb"}
+{"name":"BLAKE3_INLINE::normal","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":701,"output_sha256":"99e2e1b4012af64bab7480730eba9c84cc54b9e78a6c82928b7f4808726b877d"}
+{"name":"BLAKE3_INLINE::compressed","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":701,"output_sha256":"7afc84a09cf52d3ece43933136ae1138374cbe2c864d6c71595497e2c591a342"}
+{"name":"BLAKE3_INLINE::rd_zero","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":701,"output_sha256":"8b252a3f4303ea9c277fb5d3bfa4a8cbb832e9ea31fed64faf088c760053a6c7"}
+{"name":"BLAKE3_INLINE::aliased_operands","inline_name":"BLAKE3_INLINE","opcode":11,"funct3":0,"funct7":3,"extension":"Blake3","row_count":701,"output_sha256":"505011112985d701482bffb5b9c4e0e34c7e6800d561c9ff9162aa1a5810934e"}
 {"name":"BLAKE3_KEYED64_INLINE::normal","inline_name":"BLAKE3_KEYED64_INLINE","opcode":11,"funct3":1,"funct7":3,"extension":"Blake3","row_count":658,"output_sha256":"866c3b1e73bd091c409d1ec5601ba7cfdf1b9133624bd56c7a5afc758331e642"}
 {"name":"BLAKE3_KEYED64_INLINE::compressed","inline_name":"BLAKE3_KEYED64_INLINE","opcode":11,"funct3":1,"funct7":3,"extension":"Blake3","row_count":658,"output_sha256":"853eb9a719a96562a2afbddcf1feae303bcf10600f3499c1246c83f6e5f33ce2"}
 {"name":"BLAKE3_KEYED64_INLINE::rd_zero","inline_name":"BLAKE3_KEYED64_INLINE","opcode":11,"funct3":1,"funct7":3,"extension":"Blake3","row_count":658,"output_sha256":"7c1c5f7bdc24bb1042ae0b21c5ab3f9f2477d95327348296ba56245f2b39cc8c"}
